sync: Stop using SyncOpBase::Record for events
